Create README.md
# Food-Vader-Group-Project-1
Proposal: 
1. Project title: Comparing Restaurants Using APIs
2. Team members: Mark, Chander, Fernando, Jeanette
3. Project Description/Outline: We will be utilizing and comparing Yelp, Google, and Zomato restaurant ratings to identify the top restaurants in various parts of Texas based on searches.
4. Research Questions to answer: What are the top requested restaurants in Austin? 
What are their average ratings? Does the Vader sentiment match the yelp and google ratings? 
What could be some factors affecting this? Is it a good idea to start a business in these areas?  
What is the average, for example, Mexican restaurant rating in a specific zip code? 
We will pull average Mexican restaurant rating for each zip code, and then sort zip codes by the highest ratings to find the best restaurant in the area.
5. Data sources or Data sets to be use: http://federalgovernmentzipcodes.us/ (Fed gov zipcodes)
6. APIs to be consumed: Yelp API, Google API, Zomato API
7. Rough Breakdowns of tasks:
       Team lead: Jeanette
       Data engineer: Chander
       Tester: Mark
       Scrum Master: (All)